requirements.txt
discord.py[voice]>=2.3.2
PyNaCl>=1.5.0